Changes related to debugging MITM and other stuff
1. MITM is blocking on opera 11. It's OK with opera 10.60.
Changed some stuff, but opera 11 is still blocking, due to SSL_accept()
problem between proxy and browser. Some SSL related stuff is still
not happy with opera.

However, Firefox and Chrome are not that picky.

I can use opera 10.60 for the presentation for now.

2. MITM is caching the public key and private key now.

Don't have to read it every time. More efficient.

3. HTTPRequest.cpp is changed, to modify "User-Agent"